Fix crashes on 64-bit due to unintended address truncation on ptr vars
git-svn-id: https://svn.eduke32.com/eduke32@2005 1a8010ca-5511-0410-912e-c29ae57300e0

+++ polymer/eduke32/source/gamevars.c
@@ -422,7 +422,7 @@ int32_t Gv_NewArray(const char *pszLabel, int32_t asize)
     return 1;
 }
 
-int32_t Gv_NewVar(const char *pszLabel, int32_t lValue, uint32_t dwFlags)
+int32_t Gv_NewVar(const char *pszLabel, intptr_t lValue, uint32_t dwFlags)
 {
     int32_t i, j;
